feat: first iteration of dashboard with interactive map
I got the code example from the starter kit working. Joris was right,
we just had to fix the paths. In the future, we'll replace the json
files with API calls. I ended spending the most time fiddling with
tailwind to get to look like what I want.
The frontend is going to need a lot of work...