feat(api): make the map use api endpoint for retrieving POIs
Pretty neat how easy this ended up being. Major credits to @Wessel for
making a good start on the backend design. Figuring out api
authentication's going to be more challenging though. After that it's
just going to be more of the same, to get all the crud endpoints
working, with webforms in place.