refactor(model): delete `IncidentReport`, as it has no table
As I was working on them yesterday, I realised it didn't make a lot of
sense to put this in a separate table, if there's just a one-to-one
relation, and they'll never be used separately from each other. Instead,
I added a nullable `report` text column to incidents. If we have time to
implement a report writing feature, it can just update that column
instead.